api/types/swarm: fix EndpointVirtualIP.Addr type
The value returned from the Swarm API (and Moby <=v28) is a CIDR prefix
string. Elsewhere in the code we assert that they are CIDR prefix
strings. But it's incorrectly parsed and represented as an address in
one place in the daemon and api, respectively. Change them to a
netip.Prefix so as not to break compatibility.

Signed-off-by: Cory Snider <csnider@mirantis.com>