api/docs: remove BuildCache.Parent field for API v1.42 and up
The BuildCache.Parent field was removed in API v1.42 in [moby@e0db820].
While we had to keep the Go struct field around to backfil the field for
older API versions, it's no longer part of API v1.42 and up (using the
"omitempty" is just an implementation detail).

This patch corrects the swagger files to match this.

[moby@e0db820]: https://github.com/moby/moby/commit/e0db8207f3b84d030a71f41db70fc7fbf0535b9f

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>