api/types/sytem: remove deprecated DiskUsage.BuilderSize
The `DiskUsage.BuilderSize` field was added as part of the, then experimental,
BuildKit builder in [moby@5c3d2d5] (API v1.31). It was deprecated in API v1.32
(through [moby@b225258]) but that change still returned the field. Commit
[moby@31348af] removed it in API v1.42. This field was never documented,
and part of an experimental feature, so we can remove it altogether.

[moby@5c3d2d5]: https://github.com/moby/moby/commit/5c3d2d552b0430672d5f481ab2d37036f6e92166
[moby@b225258]: https://github.com/moby/moby/commit/b225258496f3ec8b6906b3fc6a3a2c76e50502d9
[moby@31348af]: https://github.com/moby/moby/commit/31348afa194eaff0ef9639a0c5aefa6bce4ef4b1

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>