api/types/image: remove deprecated Summary.VirtualSize field
The `VirtualSize` field became redundant with the introduction of content-
addressable images in docker v1.10 (4352da7803d182a6013a5238ce20a7c749db979a),
after which the field was identical to the `Size` field. The field was marked
to be deprecated, which happened in 1261fe69a3586bb102182aa885197822419c768c
(API v1.43) and removed in API v1.44 (913b0f51cab18a56247a950f5f1e75ca79b63039).

Now that we stop supporting API versions older than v1.44, we can drop this
field; it's a minor breakage of old API versions, but the same information
is available in the "Size" field.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>