daemon: Daemon.ContainerStats: combine some conditions and remove error
When streaming stats, the `OneShot` option is implied; streaming results
never populate the Pre* fields on the first result, so instead of producing
an error, we may as well ignore the option.

While updating this, also combine some of the conditions for non-streaming
results in an attempt to make the logic (early return on non-streaming,
one-shot requests) slightly clearer.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>