go.mod: remove "exclude" rules
These unfortunately prevent "go install github.com/moby/moby/v2/cmd/dockerd@XXX"
from working. Let's see if that works if we remove these. We may want to
add them back for the GA release of v29, but OTOH, nobody should have the
v2 module as dependency, so maybe it's not the end of the world.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>