daemon: restore: register containers without rwlayer
Under some undetermined circumstances, a container might end up with no
rwlayer on daemon restart. The restore logic would add them to the
containersReplica list but not to the 'containers' list. Thus, users
would see them on `docker ps -a`, but would be unable to remove them.

To fix that, add these containers to the 'containers' list as well.

Signed-off-by: Albin Kerouanton <albin.kerouanton@docker.com>