hack/vendor/dropreplace: Make the auto-ref pick more reliable
dropreplace now accepts an optional argument that allows to manually
specify a git ref to use for the modules.

It's not required though and if not present it will default to a master
branch from the upstream (preferred) or origin remote.

This is much more reliable than "module@master" which for some reason
tends to not resolve to the actual latest commit on the master branch.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>