Makefile: remove unused DOCKER_BUILD_OPTS
This option was added in e6d7df2e5d313800414b955e10a26d6687e7a1bf to allow
skipping copying the source code as part of `docker build` if a bind-mount
was used to improve the build performance.

This trick was replaced in 1c82d1190888f05836c550c85e4e8daedae2a813, which
now uses a separate target.

With the above, the `DOCKER_BUILD_OPTS` is no longer used, so we can remove it.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>