Makefile: remove unused DOCKER_BUILD_PKGS
This option was added in 0bff2515f74fc6f0be9f019cc8b0f08b50e36228 to control
what packages were built for the deb and rpm packages.

This repository no longer builds the deb and rpm packages, and this variable
is no longer used, so we can remove it.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>